feat: ft-wave1-repeater-reject-accept-005 - Consume deletion batch and close verification gates
